Replace old dropdown logic with new one
