+++ packages/ui/src/components/index.ts
@@ -40,7 +40,6 @@ export * from './va-date-input'
 export * from './va-date-picker'
 export * from './va-divider'
 export * from './va-dropdown'
-export * from './va-dropdown-new'
 export * from './va-form'
 export * from './va-hover'
 export * from './va-icon'
