--- packages/ui/src/components/va-dropdown-new/_variables.scss
@@ -1,6 +1,0 @@
-:root,
-:host {
-  --va-dropdown-line-height: 1;
-  --va-dropdown-content-wrapper-z-index: var(--va-z-index-teleport-overlay, 9);
-  --va-dropdown-display: inline-flex;
-}
