--- packages/ui/src/components/va-dropdown-new/index.ts
@@ -1,4 +1,0 @@
-import withConfigTransport from '../../services/config-transport/withConfigTransport'
-import _VaDropdownNew from './VaDropdownNew.vue'
-
-export const VaDropdownNew = withConfigTransport(_VaDropdownNew)
